Fix overlapping links on contact section of Worldwide Offices
The narrow column design of the Worldwide Office page means that there is an
issue where long links (that don't contain wrapping characters like `-`) will
overflow.

This is a pre-existing issue from the Whitehall app.

As a quick fix, we can allow wrapping within words. Whilst the break points are
arbitrary, this is preferable to the overlapping we currently see. Both the
`word-wrap` and `overflow-wrap` declarations are needed to support all browsers.