Ensure `nil` sponsoring organisations are handled appropriately
As part of moving the rendering of the Worldwide Office pages to
government-frontend, we republished all of the Worldwide Organisations.

However, there was an issue where some Organisations were not republished. This
meant that we saw errors in production where we were trying to access the
details of a sponsoring organisation which was nil.

This makes sure that we are correctly handling the case where there are no
sponsoring organisations linked to the office's organisation.

`worldwide_organisation` is made private as it's not referred to outside of
this class.