Fix incorrect test assertion
The assertion is passing by coincidence, because the item is nil in the
example in publishing-api.

I only noticed it thanks to the deprecation warning from minitest:

```
HowGovernmentWorksPresenterTest
DEPRECATED: Use assert_nil if expecting nil from
test/presenters/how_government_works_presenter_test.rb:14. This will
fail in Minitest 6.
```

I will also open a pull request to add a `current_prime_minister` link
to the example, so the test is more meaningful.