Ensure we are always testing all travel advice schemas
We had a flakey test which was failing since it was randomly selecting a travel advice schema to test against, and one of these example schemas actually isn't able to be rendered at the moment (fix coming). This ensures that we always test all example schemas each test run, to avoid introducing behaviour that breaks rendering for a certain case, and this not being immediately exposed.