Fix bug when rendering travel advice with no parts
In ouor rendering, we were implicitly assuming that all travel advice content items had at least one part, so when rendering content items where this is not the case we were hitting errors. This has not caused known issues on our live site as far as we're aware, but it did cause a flakey test failure since previously we were randomly selecting an example content item to render in test, and so when the selected example was the 'no-parts' travel advice example, the test was failing.

This was actually highlighting a genuine issue in our rendering. Although there are no current examples of travel advice without parts, both our schemas and travel advice publisher permits it, so the example schema represents a valid possibility within our systems.

This commit therefore ensures that any rendering relevant to parts is only executed when parts exist