Render Worldwide Corporate Information Pages
We have added a new content item to represent Corporate Information Pages that
are associated with Worldwide Organisations
(https://github.com/alphagov/publishing-api/pull/2399).

Because the Worldwide Corporate Information Pages have a link to their
Worldwide Organisation, much of the rendering code added for Worldwide Offices
and Worldwide Organisations works with this page.