Render RTL text for worldwide organisation pages
The worldwide organisation pages need the `direction-rtl` class in order to
correctly render languages with RTL writing systems like Arabic.

The groundwork is already in place for this, we simply need to include the page
type in `application.css`.

Visually, the page is now the same as it was in Whitehall (before the rendering
was moved over), except for the social media logos, which are now on the right.
This change is enabled by the components library and is desirable.

The `padding-inline-start` change ensures that the list of corporate
information pages does not have additional padding when rendered RTL.