Added presenter for new call for evidence document type
Call for Evidence is a new content type to avoid confusion for users and legal challenges. User research found that the distinction between consultation and call for evidence isn’t clear, both for departments and users. Departments are currently using the consultation content type for calls for evidence, but this can create problems and legal challenges because they are potentially obliged to publish the outcome of all consultations, whereas that’s not the case for calls to evidence.