Added default language locale for new call for evidence document type
Ideally we would simplify some of these locale strings by introducing duplication for many of the shorter strings rather than using interpolation frequently, as the current solution is difficult to understand and makes problematic assumptions about language structure. However, we are leaving these as is for now as we are not supporting rendering calls for evidence in languages other than English at present, and we want to maintain consistency with the consultation rendering for now.