Add rendering logic for the new call for evidence document type
The call for evidence document type is very similar to the consultation type. There are two differences:

- Publishers cannot attach public feedback information to the call for evidence (though this could be included in the outcome)
- The outcome of the call for evidence is not considered "final" as further calls for evidence or consultations may be conducted on a subject

We explored the possibility of doing without the legacy attachment rendering for calls for evidence, but given Whitehall still publishes attachments as pre-rendered HTML and the attachments partial uses that by default, we have chosen not to do that as part of this commit and will look into migrating all document types to the new rendering behaviour at a later date.