Fix misaligned related content sidebar
Adjustments in previous commits #2837 and #2840 did not account for
aligning the grey border with the blue border of the related content.
This small fix ensures that both columns line up.

Note that this is a quick fix and further work should be done on
these pages to consolidate the template.

See Trello for more info:

https://trello.com/c/H9MfuTul/1916-consolidate-traveladvice-and-guide-pages-in-governmentfrontend-l