Filter people roles by organisation
People linked to a worldwide organisation may have multiple roles, some of
which are related to other organsaitions. Currently, we are incorrectly
displaying these roles on the worldwide organisation pages.

This copies the approach taken in alphagov/collections#1719. We link to all
roles from the worldwide organisation, then filter out the incorrect ones in
the rendering app.

See also alphagov/publishing-api#2478 and alphagov/whitehall#8162.