Remove HTTP feature flags
The rendering of recommended related links was conditional on a feature flag,
which is set via a header within our Fastly VCL.

As an outcome of RFC-163, we are trying to simplify our CDN configuration by
removing functionality that’s no longer used, or by moving logic to other parts
of the stack.

This change was suggested in a comment on RFC-163:

  "I think the flag for Whitehall recommended links should be removed. It was put
  in as a safety mechanism for the introduction of the links, in case something
  really inappropriate was found. There are ways to manually override the links
  now, and we've never used the feature flag."

Related content is component is defined in govuk_publishing_components gem and
this is where the overriding is happening (see:
https://github.com/alphagov/govuk_publishing_components/blob/5d713ff637de14bec6abc784c35ee63a0cbc2c75/lib/govuk_publishing_components/presenters/content_item.rb#L70-L74)

It's now recommended to create feature flags as environment variables in helm
charts. See the docs:
https://github.com/alphagov/govuk-developer-docs/blob/33f1f0434aa988419df0687a829367e3f769816f/source/kubernetes/manage-app/set-env-var/index.html.md?plain=1#L7-L21

The value can then be read in Rails config:

```
config.unreleased_features = ENV["UNRELEASED_FEATURES"].present?
```