Remove legacy code to render travel alerts
Remove code from the view that would have rendered travel alerts as part of govspeak. To our knowledge this code has never been used as it wasn't possible to set alert_status in the editor until recently.

Also remove formatting from the presenter which was customised for rendering the alerts in govspeak by flattening the array of alert statuses and wrapping each in a paragraph tag. This is in readiness for rendering the alerts with warning text in the next commit.