Remove link tracking from certain 'See all updates' links
Some of these links trigger a select_content event, as they open an accordion and jump to a heading on the page as opposed to navigating to a new link. Therefore with this tracking enabled, it was sending a navigation and select_content event on click. By removing this, it will only send a select_content event on click when the accordion is expanded by the click.