AB test - replace some content on UTR page
HMRC would like to see whether a link to a video performs better than
some descriptive help text. We've agreed to run an AB test to compare
these variants.

There's no elegant way to AB test bits of content in GOV.UK, so we've
had to resort to the fragile and ugly approach of find / replacing the
content.

As per the comments in the code, this will break if the content in the
content item changes, and we'll have to update the hardcoded replacement
if the content in the B variant needs to change.

Nevertheless, we think this is a pragmatic thing to do as a one off, as
it keeps an immportant stakeholder happy, helps us learn about video
content (which is a strategic priority), and helps us learn about the
use case for content AB tests in mainstream content.