Use Time.zone.parse in ManualUpdatesPresenter
At the moment, it uses Date.parse which doesn't account for the difference
in timing between the timestamp (UTC) and local time (BST).

This causes issues with some updates being a day off if they were published
between 00.00 & 00.59.

Time.zone.parse parses the date correctly into the "London" timezone.
We still need to convert it back into a date though as it groups on
year, then by identical dates, so all updates for a day are grouped.