More robust replacement of AB test content
As suggested by @hannako, we can make the AB test replacement less
fragile by looking for a specific replacement placeholder in the content
item, instead of replacing a particular bit of content.

I've gone for a [mustache](https://mustache.github.io/) style
replacement string of {{ab_test_find_utr_number_video_links}}. There's a
possible future thought of doing other kinds of templating, so I think
starting to socialise this syntax is a good idea.

I've checked in the govspeak preview app, and the `{{...}}` syntax is ignored
by govspeak so there should be no issue entering it in the publishing
app.

I've pulled the replacement strings out into the translation file so
that they're easier to review / keep up to date.

I've also switched to three A / B / Z variants, so we can do splits like
5% / 5% / 90% where we're measuring the difference between the two 5%
samples and ignoring the 90% sample. This test is probably going to be
50% / 50%, so not strictly necessary, but this seems to be the idiomatic
way to do it.