Only AB test if page contains the placeholder
In the situation where this code is deployed, but the content item
doesn't yet contain the placeholder to be replaced, we don't want to do
any of the AB testing code (i.e. setting / reading of cookies, sending
events to google analytics).

This means this code can safely be deployed to production, and the AB
test will only begin once the magic placeholder {{ab_test_find_utr_number_video_links}}
appears in the content item.