Change initilisation method to new method
In govuk-frontend v5, the `init` function is now called when a module is
created. This meant changing how the `initAll()` function works in
govuk_publishing_components so that it does not call `init()` when
initialising components JavaScript. It also meant moving `init`
functions in components to the constructor to bring them in line with
the new way of initialising.