Use formatted name for worldwide organisations
We are currently presenting the Worldwide Organisation's title in the
logo. This means there are no line breaks in the places specified by the
user.

Updating to use the formatted title, which the publisher would have
expected since they entered this in Whitehall.

This also applies the change to Worldwide Offices and Worldwide
Corporate Information Pages, which include the Worldwide Organisation's
logo in the header.