WIP - Move AB test configuration into YAML
This is an idea / illustration of an approach for doing AB tests of
content in a slightly more maintainable way.

I still really don't like having them hard-configured in the codebase,
so I'm not actually in favour of merging this.

What I'm trying to show is that if we provide government-frontend with
the right bits of config, it's capable of dynamically adding the AB
tests needed for a particular content item. If we can do that with a
YAML file, then we could potentially do it with information on the
content item itself.

I'm imagining a world where we have something like:

    {
      ...
        "first_published_at": "2016-02-26T11:53:52Z",
        "links": {
          "ab_tests": [
            {
              "name": "FindUtrNumberVideoLinks",
              "dimension": 61,
              "base_paths": ["/find-utr-number"],
              "variants": [
                ...

in the content item, and the code works out everything from that.

We could also have something listening to the publishing queue and
configuring the Fastly dictionaries whenever this linked item updated.

If this idea is popular, I'll write it up as an RFC.