Use govuk_account_session header to toggle email signup route
Document collection email subscriptions are complicated[1]

The expected behaviour is:

If a document collection has a taxonomy topic email override in its links,
render the signup link component, and subscribe the user to the taxonomy topic instead of
to the document collection itself.

In all other cases, render a single page notification button.

The button is a gem component, and it uses the value of the attribute
`skip_account` to determine the form action of this button[2]

- If the user has an active govuk-account session, the button should post to
email-alert-frontend's /email/subscriptions/single-page/new endpoint. This endpoint enforces
the gov.uk account[3].

- If the user is not logged into their account (or they don't have an account) then the button
should post to email-alert-api's magic link endpoint, /email-signup instead. This endpoint does
not enforce the govuk account.

Current behaviour:

Prior to this commit, we were passing skip_account = "true" to the button component on all
document collection pages, and therefore hardcoding the button to post to the account free
endpoint regardless of whether or not they were logged into their account.

New behaviour:

What this commit does, is to only set the value of skip_account to "true" for users without an active
govuk-account session.

Why are we making this change?

For users with an existing subscription returning to a document collection page, the single page notification
button doesn't allow them to unsubscribe. The button text is "Stop getting emails about this page", but instead
of posting to email/subscriptions/single-page/new, which supports both subscribing AND unsubscribing[4] it is
hardcoded to post to /email-signup which supports sign up only. This means they must log into their govuk accounts
in order to manage their subscriptions.

[1] https://docs.publishing.service.gov.uk/manual/document-collection-emails-a-special-case.html
[2] https://github.com/alphagov/govuk_publishing_components/blob/main/lib/govuk_publishing_components/presenters/single_page_notification_button_helper.rb#L58-L64
[3] https://github.com/alphagov/email-alert-frontend/blob/main/app/controllers/govuk_account_signups_controller.rb#L15-L17
[4] https://github.com/alphagov/email-alert-frontend/blob/main/app/controllers/govuk_account_signups_controller.rb#L27-L34