Use edition links for worldwide organisation offices
We are switching worldwide organisations to use edition links which do
not support multi-level link expanstion.

Therefore we need to switch the rendering code to use the new link
fields that have been added, in the absence of multi-level links.

This change adds support for edition links, but also retains
compatibility with non-edition links. This will allow us to deploy the
frontend changes in advance of republishing the content items,
preventing any downtime of these pages. The code for backward
compatibility will be removed once the content items have all been
republished with edition links.