Add support for worldwide offices embedded in worldwide organisations
As part of the switch to Worldwide Organisation content items containing
the content for their offices, we need to switch the rendering of the
office pages to use these parts.

This makes that switch in a backward compatible way, so we can continue
rendering office pages under the existing content items until the office
routes are updated to point to the organisation's content item.

In a later PR, we will remove `WorldwideOfficePresenter` and it's
associated view.