Update NewsImage to render multiple defaults
The logic to render an image for a news item is:

- render the details > image if there is one
- render the organisation > default image if there is one
- render a placeholder

Organisation here could be one of two types:

- Organisation, presented in the primary_publishing_organisation link
- WorldwideOrganisation, presented in the worldwide_organisations link

The placeholder is also dependent on which type of document and so which
type of organisation the content links to.

Here we allow the module to render the default image from the worldwide
organisation link, when the document is the appropriate type and one is
available.

Right now this change will not be used as Whitehall currently sets
details > image to the default image for worldwide organisation news
items, but this is set to change as we want this logic out of the
publishing applications and for all related concepts like 'the default
news image' to work the same way.

With this change, once we do turn the replacement in Whitehall off,
everything will continnue to work without change.