Use publishing component to render contents list on service manual guide
When the contents list on service manual guides was being read out by
screen readers it would read the dashes in the contents list as "en dash"
This was not desired behaviour and has been fixed in the
govuk-publishing-components. This uses the fixed publishing component to
take advantage of the fixed behaviour.

Using the new publishing component also makes the page more consistent
with other pages within the app.