Use relative URL for worldwide organisation links
Links from the Worldwide Organisation pages currently use the `web_url`,
which is absolute, and never includes the draft part of the URL.

This means that a user viewing a draft Worldwide Organisation will be
taken to the live version of the page.

Updating to make the URL relative, meaning users are always taken to the
correct version of the page.

This was never previously an issue, since worldwide organisations have
not previously been editionable.