Make contact_for_office tolerant of missing data
Worldwide organisation content items should have both
office_contact_associations and a link to contacts.

If they don't have both, we can either error (what the code used to do),
or just skip the contact section (what this commit does).

Note that content_item is actually a GdsApi Response, not a hash.
GdsApi's Response class delegates `dig` through its `to_hash` method,
but not `fetch`. We explicitly turn it into a hash instead so we can use
fetch (which will error if details or links is missing, but return a
default [] if office_contact_associations or contacts is missing).