Display GOV.UK Chat promo on certain pages
We want to show a promo to the new GOV.UK Chat application on certain pages
on GOV.UK.

The list of URLs is quite short and cover only `answer` and `guide`
content types.

The CTA is to be shown in the sidebar on the right of the relevant
pages.