Replace organisation logos with metadata
This commit is a spike to explore how easy/hard it would be to replace
organisation logos on HTML attachments with a call to the metadata component
instead:
https://components.publishing.service.gov.uk/component-guide/metadata#on_a_dark_background_without_padding

Some organisations have custom logos, but we only display the logo if it will be
the only logo on the page. If it will sit alongside other logos, we hide the custom
logo. This behaviour is weird and unexpected.