Fix main-content-container to work with govuk-frontend columns
This commit removes some of the rules that were previously set for the
`.main-content-container` and instead uses the column rules from
`govuk-frontend`.

The container has now been renamed as `contents-container` to be
consistent with the `contents-list-container`. I had hoped to remove
this class entirely, but it's required to apply different float rules
when in RTL reading mode.

This also adds the new class for collapsing columns to full width
when printing.