Use GraphQL for rendering news articles
This retrieves the content from Publishing API, instead of Content
Store, using a GraphQL query.

Note this has been done in the way that reduces the number of changes
needed in this application. As a result, we have introduced a small
"hack" to return the data in `GdsApi::Response` format but with this app
rewriting the hash keys to make them all underscore based.