Restore GOV.UK Chat banner to nested guide page
In https://github.com/alphagov/government-frontend/pull/3440 we changed
the mechanism for rendering the GOV.UK Chat banner to utilise request
path rather than content item base_path. A side effect of this was that
for any existing guides that we were embedding the banner on lost the
banner on any of their nested sections.

This commit adds each individual path of guides that previously had the
banner.