Hide email button if content not in English
GOV.UK does not support email subscriptions to foreign language content.
The single page notification button should only have been added to english
pages, but was added to all pages. The consequence has been users signing
up to an email subscription that will never match a content change[1]

We could make the button work so that clicking the button on eg a welsh page results
in the user signing up to the equivalent content in English. But we decided against this:

1. There is no indication that there is a user need for that behaviour
2. The code to fetch the english base-path would be added to the button component in the
gem. The logic for this button is already very complicated, and adding further complexity
feels like a bad idea.
3. Not all our publishing applications create foreign content with a link back to the english
version. So if the button was added to other content types in the future it wouldn't work as
expected eg https://www.gov.uk/api/content/budd-dal-plant-16-19
4. Hiding the signup button is the same approach used elsewhere [2]

[1] https://github.com/alphagov/email-alert-service/blob/main/email_alert_service/models/major_change_message_processor.rb#L23-L26
[2] https://github.com/alphagov/collections/blob/ea307473313aa5138e6db6dd16eb26f1bf453bfd/app/views/world_location_news/show.html.erb#L56-L86