Memoize manual section presenter
We're currently calling the main method, and then looping over the
result. In each loop, we call the method again so that we can get the
length. See manual_section.html.erb, which does roughly:

   @content_item.main.each.with_index do |item, index|
     # ...
     {
       data_attributes: {
         ga4_event: {
           # ...
           index: index,
           index_total: @content_item.main.length,

... the effect of this is that we run the main method n times, where n
is the number of <h2> elements in the body. This results in us spending
a lot of time parsing the same document and running the same CSS / XPath
selectors on it, which adds up to a lot of time.

Memoizing the result of the main method shaves several seconds off the
time to render www.gov.uk/guidance/complete-the-school-census/data-items-2024-to-2025
on my machine.

I did look at some other optimizations (e.g. traversing the nokogiri
document manually, instead of using css / xpaths), but once it's
memoized those only save microseconds so it's not worth it.