Add document_type and schema_name to prom labels
This results in metrics like:

    http_requests_total{action="show",controller="content_items",document_type="answer",schema_name="answer",status="200"} 1

This will allow us to investigate the performance of different content
types, for example how long does it take to load a manual vs. an answer
or a case_study. Currently we can't distinguish these in the metrics for
government-frontend, because it uses the same controller action for all
content types.

If there are some content types which take a much longer time to load,
maybe there are performance issues we should address.

This follows on from https://github.com/alphagov/govuk_app_config/pull/424
which adds the functionality to support setting prometheus labels in this way.