Rename aside to secondary-text
Calling the secondary content block 'aside' doesn't reflect that it has standard positioning now, it might be a reference to the visuals of a previous design. Rename it to 'secondary-text' which is more descriptive of the fact that it's visually the same as the main text but is also supplementary to it and optional. 'secondary-text' is also more reflective of the CSS class 'app-c-banner__text' that wraps it.