Replace spacing override class in templates with banner component class
Instead of using a generic override to set styles within the banner component, use the component's own text style class. This makes it clear that we're relying on the styles on this content to be the same as those of the component, and ensures they stay consistent if the component is updated in the future.