Make banner component require both title and text
This component now requires title and text attribute values passed to it, otherwise it will not render. This reflects how the component is currently used. If the component needed to become more flexible and work for more use cases, we should consider moving it to the gem.

(We're generally trying to reduce complexity in apps, instead of adding further unnecessary options; there's also been some discussion about how the gem inverse header component could in the future be made more flexible and replace this banner component so we want to make that easier where possible instead of allowing more use cases and potentially bloating the app component further.)