Add rendering for Sub Facet value and sub facet link
Any facet that is a sub facet needs to be rendered with a label in the format "<main facet> - <sub facet>". The link also needs to include the main facet as part of query params in addition to the sub facet. The linked main facet label and value is passed through in the finder schema via allowed values as illustrated in the test.