Create a temporary alternative to interpolation as it doesn't work for Welsh
The existing interpolation of "is being" or "was" into the sentence
stored in another_website_html didn't work for Welsh based on the translations
obtained today. So as a temporary measure, I've created a key that contains
the full sentence with only the url being interpolated.

This means the call_for_evidence template will no longer be fetching the
:another_website_html which is populated for many languages (unusually!)
and we only support the new keys :open_another_website_html in Welsh and English.

Currently there are 800+ calls for evidence, and only one is in cy, so this feels
low risk. But we will come back and address this tech debt asap - this is just to
meet an urgent request.

Edition.where(type: "CallForEvidence").pluck(:primary_locale).uniq
=> ["en", "cy"]