Replace detailed_guide with case_study schema
- As we are removing detailed_guide from government-frontend, I have updated the schema back to case_study as case_study is not yet removed from government-frontend even though it had been moved over to frontend.
But once case_study is also removed from government-frontend, we need to update the schema with the relevant document types that are still being rendered from government-frontend.
Also we would have to remove or update the test for - correct meta tags are displayed for pages with images - once case_study is removed in this line - https://github.com/alphagov/government-frontend/blob/6371f6ef101f3c2248f57085ef7c868fb963c3a3/test/integration/meta_tags_test.rb#L49
This is so because the meta[name='twitter:image'] is displayed for documents that have `image` attribute in `details` hash. As of now we have migrated speech,specialist_document and case_study to frontend already and the tests are already in place in frontend.
https://github.com/alphagov/frontend/blob/5ebabc8d1fd547b210be65be7e2877682e961339/spec/support/meta_tags.rb