Update helper method to use available translations for "on"
We should use the translation when available, and will fallback
to the English default if there isn't one.

This helper method is only used by Consultation pages (call for
evidence pages are now rendering from Frontend and the code in
government-frontend will be removed soon). We will also be
migrating consultation pages soon.