Increase h1 font-size to fix accessibility issue
There are two h1 elements on the page, one includes the `content_title` and the other includes the `content_part_title`.

When different font sizes are used for each h1 element, it implies a structure and relationship that may not be reliably perceived by some users of assistive technology.

This change ensures that both h1 elements use the same font-size to fix the accessibility issue, Level A failure of WCAG 2.2 Success Criteria 1.3.1 Info and Relationships.